algorithm file
Setup
For this task, we'll need:
Python: To run our script
Pip: Necessary to install Python packages
Now we can install some packages using pip, open your terminal and type these out.

pip install numpy
pip install sklearn
Numpy: Useful mathematical functions

Sklearn: Machine learning tools for python

The closeness/ proximity amongst samples of data determines their neighborhood. There are several ways to calculate the proximity/distance between data points depending on the problem to solve. Most familiar and popular is the straight-line distance (Euclidean Distance).

Generally, neighbors share similar characteristics and behavior that's why they can be treated as they belong to the same group.

This is the main idea of this simple supervised learning classification algorithm.

Now, for the K in KNN algorithm that is we consider the K-Nearest Neighbors of the unknown data we want to classify and assign it the group appearing majorly in those K neighbors. For K=1, the unknown/unlabeled data will be assigned to the class of its closest neighbor.

Consider a document containing 100 words wherein the word ‘car’ appears 7 times.
The term frequency (tf) for phone is then (7 / 100) = 0.07. Now, assume we have 1 million documents and the word car appears in one thousand of these. Then, the inverse document frequency (i.e., IDF) is calculated as log(10,00,000 / 100) = 4. Thus, the Tf-IDF weight is the product of these quantities: 0.07 * 4 = 0.28.

Generally, the value of K is taken to be as √n, where n = number of data samples. We need to tune in the K parameter based on the validation set. The value of K will smooth out the boundaries between classes. 
The overhead of calculating distances for every data whenever we want to predict is really costly. So, K-NN is not useful in real-time prediction.